Make OGR_VRT driver optional and buildable as plugin
